Final commit

+++ include/taylor_sine.h
@@ -1,7 +1,7 @@
 #ifndef TAYLOR_SINE
 #define TAYLOR_SINE
 
-long factorial(int in);
+double factorial(int in);
 double taylor_sine(double x, int n);
 
 #endif
